Add smoothing module
