+++ pyaldata/__init__.py
@@ -7,3 +7,4 @@ from .movement_onset_and_peak import *
 from .data_cleaning import *
 from .io import *
 from .cmp import *
+from .smoothing import *
